backend/pyproject.toml
[project]
name = "temp-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "bcrypt>=5.0.0",
    "flask>=3.1.2",
    "flask-cors>=4.0.0",
    "pyserial>=3.5",
    "PyJWT>=2.8.0",
]
